Break out points tests to their own file
